fastboot: initial implementation of a local fastboot utility
the idea is to flash the .img file locally, e.g. from the target. Instead of
using the rawprogram.xml file format, we would do as-if fastboot runs from the
target.

fastboot.c is derived from aboot.c file in LK, it implements a simple parsing of
the sparse image and write to the file. It has been tested with sparse images
generated with make_ext4fs as well as images generated with mkfs.ext4+img2simg.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>