dragonboard410c: add aosp flavor
Some community users want to build AOSP for DB410c, using the Landing Team
kernel (or mainline + patches). For this specific use case, we want to have a
bootloader package that provide the Android partition table, but uses the
firmware (sbl, tz, rpm, LK, ..) which are typically used in Linux builds.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>