dragonboard410c: fix up android partition definitions
For the last partition it is a much better idea to leave the size as 0, to let
the script fill up the remaining space. In any case, the current values that we
had were 'too big' for the onboard eMMC on DB410c, and trying to flash
gpt_both0.bin would result in:

[760860] [760860] Attempt to write partition image.
[760860] [760860] Writing GPT partition
[760860] [760860] Invalid last_usable_lba
[760870] [760870] GPT: Backup signature invalid cannot write GPT

This patch fixes this issue, and LK will anyways always resize the last
partition to match the emmc size.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>